runtime/pyproject.toml
[tool.poetry]
name = "samplifier-backend-api"
version = "0.1.0"
description = "Torpedo Empresa web api"
authors = ["Marcio Miranda <marcio.miranda@druid.com.br>"]

[tool.poetry.dependencies]
python = "^3.7"
aws-lambda-powertools = "^1.20.2"
requests = "^2.25.1"
druid-libs = "^0.1.15"
aws-requests-auth = "^0.4.3"
boto3 = "^1.17.0"
PyJWT = "2.0.0" # Newer versions are incompatible, do not update
chevron = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-cov = "2.12.0"
freezegun = "1.1.0"
black = "^20.8b1"
flake8 = "^3.9.0"
isort = "^5.8.0"
python-mock = "^0.0.3"
requests-mock = "^1.9.3"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120
skip = "example"

[tool.black]
line-length = 120
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | example
)
'''

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -vv"
testpaths = "./tests"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

